Changing all int log-levels to enum type

+++ libjob_queue/src/forward_model.c
@@ -239,7 +239,7 @@ static void forward_model_json_fprintf(const forward_model_type * forward_model,
 
   fprintf(stream, "\n\"ert_version\" : [%d, %d, \"%s\"],\n", version_get_major_ert_version(), version_get_minor_ert_version(), version_get_micro_ert_version());
   fprintf(stream, "\"simulation_id\" : \"%ld\",\n", random());
-  fprintf(stream, "\"ert_pid\" : \"%ld\"\n", getpid()); //Long is big enough to hold __pid_t
+  fprintf(stream, "\"ert_pid\" : \"%ld\"\n", (long) getpid()); //Long is big enough to hold __pid_t
   fprintf(stream, "}\n");
   fclose(stream);
   free(json_file);
